requirements.txt
dspy-ai
requests
beautifulsoup4
pandas
numpy
python-dotenv
tqdm
pytest
tenacity

